stock02/requirements.txt
line-bot-sdk
flask
bs4
lxml
